extract JSX and convert to TSX
