+++ assets/prototype/application/services/toaster/LoadingToast.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { Spinner } from '@blueprintjs/core';
+import { LoadingToastProps } from './types';
+
+const LoadingToastNotice: React.FunctionComponent<LoadingToastProps> = ({ message }): JSX.Element => {
+	return (
+		<div
+			style={{
+				display: 'flex',
+				flexFlow: 'row nowrap',
+				alignContent: 'flex-start',
+				justifyContent: 'flex-start',
+			}}
+		>
+			<Spinner size={Spinner.SIZE_SMALL} />
+			<span style={{ marginLeft: '1rem' }}>{message}</span>
+		</div>
+	);
+};
+
+export default LoadingToastNotice;
